MLBEDSW-1498: Add Resize_Bilinear operator support
This patch adds support for the ResizeBilinear operator.

It is implemented using a 2x2 Nearest Neighbor upscale
followed by a 2x2 Average Pool.
Depending on the argument align_corners
the output is either of shape:
    - (2 * M, 2 * N) when align_corners == True, or
    - (2 * M - 1, 2 * N - 1) when align_corners == False
where (M, N) is the input shape.

The padding mode is SAME when align_corners == True
and VALID when align_corners == False.

The argument half_pixel_centers is out of scope and is
as of now ignored.

Note that only upscaling by a factor of 2 is supported.

Change-Id: Ia6d6d010c4f1bb13f5f839bc8d16872a626d9a3b
Signed-off-by: Dwight Lidman <dwight.lidman@arm.com>