MLBEDSW-4585: Resolve NumPy setup requirement
This commit fixes a potential end user problem where
the user needs NumPy installed in order to run setup.py.
This creates a situation where if the user needs to
manually install NumPy in order to install Vela.
This solution moves the NumPy dependent code into the
setup process and allows it to be installed as a
dependency first.

Signed-off-by: Dwight Lidman <dwight.lidman@arm.com>
Change-Id: I5d7d3d30c3755e8f7c369df8d34738f5c53877dd