MLBEDSW-4783: Fix issue with relative paths to config files
One level deep relative paths (ie ./vela.ini)  were treated as the name of a
folder in config_files was ".". They are now treated as relative paths.

The warning message when using an absolute path has also been moved to
to the error message instead for a better user experience.

Signed-off-by: Rickard Bolin <rickard.bolin@arm.com>
Change-Id: I7f7d4f904b9fbba97593e42203566057a2d36925