MLBEDSW-8117: Incorrect stride check for IFM/IFM2 and OFM
The constraint check for the IFM/IFM2/OFM strides were coded
according to an incorrect version of the specification.

Changed the check to verify that the strides are a multiple
of 16 bytes. Also changed the wording in the exception message
to clarify if it is a stride or value violating the constraint.

Test case had two stride settings violating the constraint,
after this change one of them still fails the check, so
no change to tests, except in comments clarifying what is
being tested.

Change-Id: I93815d8bb08303b5f747c947c0bbd461b12895e3
Signed-off-by: Björn Davidsson <bjoern.davidsson@arm.com>